requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.6.3
attrs==23.1.0
build==0.10.0
CacheControl==0.13.1
cachetools==4.2.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.7
crashtest==0.4.1
distlib==0.3.7
dnspython==2.4.2
dulwich==0.21.6
exceptiongroup==1.1.3
Faker==20.1.0
fastapi==0.104.1
fastapi-filter==1.0.0
filelock==3.12.4
flake8==6.1.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
isort==5.13.1
jaraco.classes==3.3.0
jsonschema==4.17.3
keyring==24.2.0
magic-filter==1.0.12
mccabe==0.7.0
more-itertools==10.1.0
motor==3.3.2
msgpack==1.0.6
multidict==6.0.4
packaging==23.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pyflakes==3.1.0
pymongo==4.6.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-telegram-bot==13.13
pytz==2023.3.post1
pywtdlib==0.2.0
PyYAML==6.0.1
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
Telethon==1.33.0
tomlkit==0.12.1
tornado==6.4
trove-classifiers==2023.9.19
typing_extensions==4.8.0
tzlocal==2.1
urllib3==2.0.5
uvicorn==0.24.0.post1
virtualenv==20.24.5
xattr==0.10.1
yarl==1.9.3
zipp==3.17.0
